+++ eq_project/__manifest__.py
@@ -21,7 +21,7 @@
 {
     'name': "Equitania Projekt",
     'license': 'AGPL-3',
-    'version': '1.0.60',
+    'version': '1.0.61',
     'category': 'project',
     'description': """Extensions for project""",
     'author': 'Equitania Software GmbH',